chore: Don't clear out block env on block exit
My initial assumption was wrong: once the block is exited, there still
could be references to its local variables and such. See the test
`closure/closed_closure_in_function.lox` for the simplest example.

Unfortunately, this fact effectively causes a reference cycle: a
function holds a closure to the environment which contains the function
itself. For now, disable leak detection to prevent tests from failing
because of that, we'll see what is possible to do with it.

Fixes:
* closure/assign_to_closure.lox
* closure/closed_closure_in_function.lox
* closure/reference_closure_multiple_times.lox
* closure/unused_later_closure.lox
* while/closure_in_body.lox

A bunch of tests leak, but have their leaks silenced now:
* closure/assign_to_shadowed_later.lox
* closure/close_over_later_variable.lox
* closure/open_closure_in_function.lox
* closure/reuse_closure_slot.lox
* closure/shadow_closure_with_local.lox
* constructor/return_in_nested_function.lox
* function/local_recursion.lox
* regression/40.lox
* this/nested_class.lox
* variable/early_bound.lox